+++ SlashBar/UI/Shell/SidePanelDockShelfWindow.xaml.cs
@@ -1,6 +1,7 @@
 using System.Collections.ObjectModel;
 using System.Windows;
 using System.Windows.Input;
+using SlashBar.Modules.Native;
 
 namespace SlashBar.UI.Shell;
 
@@ -26,8 +27,7 @@ public partial class SidePanelDockShelfWindow : Window {
     }
 
     private void PositionBottomLeft() {
-        var screen = System.Windows.Forms.Screen.PrimaryScreen
-            ?? System.Windows.Forms.Screen.AllScreens[0];
+        var screen = ScreenNative.GetBarScreen();
         var area = screen.WorkingArea;
 
         Left = area.Left;